dropdown add delete

+++ src/components/productCatalog/DropDownList/DropDownList.tsx
@@ -86,8 +86,12 @@ const DropDownList: React.FC<DropdownProps> = ({
   );
  
    const subsectiondata=useSelector((state:any)=>state.productCatalog.cuisineData.data);
+   const deleteApicall=useSelector((state:any)=>state.productCatalog.deletesubsectionsuccess);
+
 
   //  console.log("subsectiondata",subsectiondata);
+  // console.log("deleteApicall",deleteApicall);
+  
    
   const getdatafrosaga = () => {
     dispatch(fetchDropDownRequest(payload));
@@ -134,12 +138,12 @@ const DropDownList: React.FC<DropdownProps> = ({
 
   useEffect(() => {
     dispatch(fetchDropDownRequest(payload));
-  }, []);
+  }, [dropDownType]);
 
   const handleOptionMouseDown = (event: React.MouseEvent) => {
     event.stopPropagation();
     if (dropDownType) {
-      getdatafrosaga();
+      // getdatafrosaga();
     }
   };
 
@@ -210,6 +214,22 @@ const DropDownList: React.FC<DropdownProps> = ({
       setValue(name, option.name);
       trigger(name);
     }
+
+    const viewdata={
+      locationId: locationid,
+    type: SubcategoryId && "SUB_CATEGORY",
+    parentId: SubcategoryId &&SubcategoryId,
+    }
+
+    if(SubcategoryId )
+    {
+      dispatch(fetchDropDownRequest(viewdata));
+    }
+   
+    
+
+
+    
     setSearchTerm("");
   };
 
@@ -241,7 +261,24 @@ const DropDownList: React.FC<DropdownProps> = ({
       id:value,
       type:dropDownType
     }
-    dispatch(deleteDropDowRequest(deletedItem));
+
+    const viewdata={
+      locationId: locationid,
+    type: dropDownType,
+    parentId: SubcategoryId &&SubcategoryId,
+    }
+
+    if(deletedItem) {
+      dispatch(deleteDropDowRequest(deletedItem));
+
+      if(deleteApicall==="success"){
+        dispatch(fetchDropDownRequest(viewdata));
+      }
+    
+
+
+    }
+
   };
 
   // const handleBlur = () => {
@@ -250,6 +287,7 @@ const DropDownList: React.FC<DropdownProps> = ({
   // value={type === "checkbox"
   //   ? selectedOptions.map((opt) => opt.name).join(", ")
   //   : selectedOptions[0]?.name || ""}
+  const [SubcategoryId,setSubCategoryId]=useState<string>("")
 
   const handleCheckboxChange = (option: Option) => {
     if (type == "checkbox") {
@@ -264,6 +302,16 @@ const DropDownList: React.FC<DropdownProps> = ({
         }
       });
     }
+    else if (type === "radio") {
+      setSelectedOptions([option]);
+      setValue(name, option.name);
+      trigger(name);
+    }
+    if(dropDownType==="CATEGORY")
+    {
+      setSubCategoryId(option.id);
+    }
+
   };
 
   const handleNewItemAdd=()=>{
@@ -274,7 +322,7 @@ const DropDownList: React.FC<DropdownProps> = ({
       locationId:locationid,
       name:newValue,
       type:dropDownType,
-      parentId:""
+      parentId: SubcategoryId &&SubcategoryId,
 
       
     }
@@ -284,12 +332,34 @@ const DropDownList: React.FC<DropdownProps> = ({
         // handleSelect(newItem);
         setSearchTerm("");
         setAddNewButton(false);
+        const viewdata={
+          locationId: locationid,
+        type: dropDownType,
+        parentId: SubcategoryId &&SubcategoryId,
+        }
+        if(addNewButton && newItem)
+        {
+          dispatch(addDropDowRequest(newItem));
+          dispatch(fetchDropDownRequest(viewdata));
+        }
 
-    dispatch(addDropDowRequest(newItem));
+ 
     console.log("option-updated",options)
 
 
   }
+  const [Loading,setLoading]=useState<boolean>();
+
+  useEffect(()=>{
+    if(!options)
+    {
+      setLoading(true);
+    }
+    else{
+      setLoading(false);
+    }
+
+  },[options])
  
   
 
@@ -325,6 +395,7 @@ const DropDownList: React.FC<DropdownProps> = ({
                 onClick={() => {
                   onToggle();
                   setShowselectedOption(true);
+                  dispatch(fetchDropDownRequest(payload));
                 }}
                 alt="dropdown"
                 className="dropdownimageclosed"
@@ -335,6 +406,7 @@ const DropDownList: React.FC<DropdownProps> = ({
                 onClick={() => {
                   onToggle();
                   setShowselectedOption(false);
+                  dispatch(fetchDropDownRequest(payload));
                 }}
                 alt="dropdown"
                 className="dropdownimageopen"
@@ -348,14 +420,19 @@ const DropDownList: React.FC<DropdownProps> = ({
         </div>
       </div>
 
-      {dropdownopen && (
+      {
+         dropdownopen && (
         <div className="dropdown-body">
           <div className="Dropdown-lists-and-edit">
             <ul
               className="dropdown-options"
               onMouseDown={handleOptionMouseDown}
             >
-              {options?.length > 0 ? (
+
+              {
+                Loading ? <div  className="dropdown-no-options">Loading</div> :(
+                  <div>
+                    {!Loading && options?.length > 0 ? (
                 options?.map((option, index) => {
                   const isOptionSelected = options.some(
                     (opt) => opt?.id === option?.id
@@ -406,9 +483,13 @@ const DropDownList: React.FC<DropdownProps> = ({
               ) : (
                 <li className="dropdown-no-options">No options found</li>
               )}
+                  </div>
+                )
+              }
+              
             </ul>
             <div className="edititem">
-              {!editList && editValues && (
+              {options?.length >0 &&!editList && editValues && (
                 <p
                   className="editiconimage"
                   onMouseDown={handleOptionMouseDown}
@@ -459,7 +540,7 @@ const DropDownList: React.FC<DropdownProps> = ({
                     </p>
                   )}
                 </div>
-                {addNew && !addNewButton && (
+                {!Loading && addNew && !addNewButton && (
                   <button
                     className="dropdown-addbutton"
                     onClick={(e) => {
@@ -474,7 +555,10 @@ const DropDownList: React.FC<DropdownProps> = ({
             </div>
           }
         </div>
-      )}
+      )
+      }
+
+    
     </div>
   );
 };
